// types/loan.types.ts
export interface LoanStats {
    total_count: number;
    active_count: number;
    completed_count: number;
    total_disbursed: number;
    total_outstanding: number;
}